Restructure files to use fragments on login/signup activity.
