+++ Atlas/android/app/src/main/java/com/bounswe2017/group10/atlas/auth/LoginFragment.java
@@ -1,0 +1,7 @@
+package com.bounswe2017.group10.atlas.auth;
+
+import android.support.v4.app.Fragment;
+
+
+public class LoginFragment extends Fragment {
+}
