+++ Atlas/android/app/src/main/java/com/bounswe2017/group10/atlas/auth/SignupFragment.java
@@ -1,0 +1,8 @@
+package com.bounswe2017.group10.atlas.auth;
+
+
+import android.support.v4.app.Fragment;
+
+
+public class SignupFragment extends Fragment {
+}
