+++ Atlas/android/app/src/main/res/layout/activity_auth.xml
